+++ main.py
@@ -4,6 +4,7 @@ from lightning.pytorch.utilities.types import TRAIN_DATALOADERS
 from monai.inferers import sliding_window_inference
 
 import nets.nets
+from monai.metrics.meandice import compute_dice
 from config.config import Config
 from preprocessing.covid_dataset import CovidDataset
 import monai.data
@@ -22,24 +23,39 @@ import lightning as L
 
 torch.set_float32_matmul_precision('high')
 
+
 class Net(L.pytorch.LightningModule):
     def __init__(self):
         super(Net, self).__init__()
         self.save_hyperparameters()
-        self.model = nets.nets.covid_unetr
-        self.dice_metric = DiceMetric(include_background=False, reduction="mean")
-        self.train_dice_metric = DiceMetric(include_background=False, reduction="mean")
-        self.loss_function = monai.losses.GeneralizedDiceLoss(sigmoid=True, include_background=False)
+        self.config = None
+
         self.post_pred = monai.transforms.Compose(
-            [monai.transforms.Activations(sigmoid=True), monai.transforms.AsDiscrete(threshold_values=0.5)])
-        self.post_label = monai.transforms.Compose([monai.transforms.AsDiscrete(threshold_values=0.5)])
+            [monai.transforms.EnsureType(data_type='tensor'), monai.transforms.Activations(sigmoid=True), monai.transforms.AsDiscrete(threshold=0.5)])
+        self.post_label = monai.transforms.Compose([monai.transforms.AsDiscrete(threshold=0.5)])
+
         self.best_val_dice = 0
         self.best_val_epoch = 0
+
         self.validation_step_outputs = []
         self.train_step_outputs = []
+
         self.training_ds = None
         self.validation_ds = None
         self.test_ds = None
+
+        self.model = None
+        self.dice_metric = None
+        self.train_dice_metric = None
+        self.loss_function = None
+
+    def configure_training_process(self):
+        self.model = self.config.get_network()
+        logging.info(f"Using model: {self.model}")
+
+        self.dice_metric = DiceMetric(include_background=True, reduction="mean", get_not_nans=False)
+        self.train_dice_metric = DiceMetric(include_background=True, reduction="mean", get_not_nans=False)
+        self.loss_function = monai.losses.DiceLoss(sigmoid=True)
 
     def forward(self, x):
         return self.model(x)
@@ -48,7 +64,7 @@ class Net(L.pytorch.LightningModule):
         # Load images and masks
         logging.info(f"Loading images from {COVID_CASES_PATH}")
         images = load_images_from_path(COVID_CASES_PATH)
-        labels = load_images_from_path(LUNG_MASKS_PATH)
+        labels = load_images_from_path(self.config.get_mask_data_path())
 
         # Convert images and masks to a list of dictionaries with keys "img" and "mask"
         data_dicts = np.array([{"img": img, "mask": mask} for img, mask in zip(images, labels)])
@@ -58,9 +74,9 @@ class Net(L.pytorch.LightningModule):
         shuffler.shuffle(data_dicts)
         data_dicts = list(data_dicts)
 
-        # Split the data into training (70%), validation (10%), and test sets (20%)
-        test_split = int(len(data_dicts) * 0.2)
-        val_split = int(len(data_dicts) * 0.1)
+        # Split the data into training (70%), validation (20%), and test sets (10%)
+        test_split = int(len(data_dicts) * 0.1)
+        val_split = int(len(data_dicts) * 0.2)
 
         train_paths = data_dicts[test_split + val_split:]
         val_paths = data_dicts[test_split:test_split + val_split]
@@ -92,44 +108,37 @@ class Net(L.pytorch.LightningModule):
         return test_dataloader
 
     def configure_optimizers(self):
-        optimizer = torch.optim.AdamW(self.model.parameters(), lr=1e-3, weight_decay=1e-5)
+        optimizer = torch.optim.Adam(self.parameters(), lr=1e-3)
         return optimizer
 
     def training_step(self, batch, batch_idx):
-        inputs, labels = batch["img"], batch["mask"]
-        outputs = self.forward(inputs)
-
-        loss = self.loss_function(outputs, labels)
-        outputs = [self.post_pred(i) for i in decollate_batch(outputs)]
+        inputs, labels = batch["img"].cuda(), batch["mask"].cuda()
+        raw_outputs = self.forward(inputs)
+        loss = self.loss_function(raw_outputs, labels)
+        self.log("train_step_loss", loss, on_step=True, on_epoch=False, prog_bar=True)
+        outputs = [self.post_pred(i) for i in decollate_batch(raw_outputs)]
         labels = [self.post_label(i) for i in decollate_batch(labels)]
-        train_dice_step = self.train_dice_metric(y_pred=outputs, y=labels)
+        self.train_dice_metric(y_pred=outputs, y=labels)
 
-        # Take the mean of the tensor before logging it
-        # train_dice_step_mean = train_dice_step.mean()
-        #
-        # self.log('train_dice', train_dice_step_mean, on_step=True, on_epoch=True, prog_bar=True, logger=True)
+        self.log("train_step_dice", self.train_dice_metric.aggregate().item(), on_step=True, on_epoch=False, prog_bar=True)
 
         train_loss_dictionary = {"loss": loss}
         self.train_step_outputs.append(train_loss_dictionary)
         return train_loss_dictionary
 
     def on_train_epoch_end(self) -> None:
-        train_loss = 0
-        for output in self.train_step_outputs:
-            train_loss += output["loss"].sum().item()
-
-        mean_train_loss = torch.tensor(train_loss / len(self.train_step_outputs)) # Total loss of batches / number of batches
+        avg_loss = torch.stack([i["loss"] for i in self.train_step_outputs]).mean()
         mean_train_dice = self.train_dice_metric.aggregate().item()
         self.train_dice_metric.reset()
 
-        self.log_dict({"train_dice": mean_train_dice, "train_loss": train_loss / len(self.train_step_outputs)}, prog_bar=True)
+        self.log_dict({"train_dice": mean_train_dice, "train_loss": avg_loss}, prog_bar=True)
 
         tensorboard_logs = {
             "train_dice": mean_train_dice,
-            "train_loss": mean_train_loss,
+            "train_loss": avg_loss,
         }
 
-        self.logger.experiment.add_scalars("losses", {"train": mean_train_loss}, self.current_epoch)
+        self.logger.experiment.add_scalars("losses", {"train": avg_loss}, self.current_epoch)
         self.logger.experiment.add_scalars("dice", {"train": mean_train_dice}, self.current_epoch)
         self.logger.log_metrics(tensorboard_logs, step=self.current_epoch)
 
@@ -146,27 +155,25 @@ class Net(L.pytorch.LightningModule):
         labels = [self.post_label(i) for i in decollate_batch(labels)]
         self.dice_metric(y_pred=outputs, y=labels)
 
-        validation_loss_dictionary = {"loss": loss}
+        self.log("val_step_dice", self.dice_metric.aggregate().item(), on_step=True, on_epoch=False, prog_bar=True)
+
+        validation_loss_dictionary = {"val_loss": loss}
         self.validation_step_outputs.append(validation_loss_dictionary)
         return validation_loss_dictionary
 
     def on_validation_epoch_end(self) -> None:
-        val_loss = 0
-        for output in self.validation_step_outputs:
-            val_loss += output["loss"].sum().item()
-
-        mean_val_loss = torch.tensor(val_loss / len(self.validation_step_outputs))
+        avg_loss = torch.stack([x["val_loss"] for x in self.validation_step_outputs]).mean()
         mean_val_dice = self.dice_metric.aggregate().item()
         self.dice_metric.reset()
 
-        self.log_dict({"val_dice": mean_val_dice, "val_loss": val_loss / len(self.validation_step_outputs)}, prog_bar=True)
+        self.log_dict({"val_dice": mean_val_dice, "val_loss": avg_loss}, prog_bar=True, on_epoch=True, on_step=False)
 
         tensorboard_logs = {
             "val_dice": mean_val_dice,
-            "val_loss": mean_val_loss,
+            "val_loss": avg_loss,
         }
 
-        self.logger.experiment.add_scalars("losses", {"val_loss": mean_val_loss}, self.current_epoch)
+        self.logger.experiment.add_scalars("losses", {"val_loss": avg_loss}, self.current_epoch)
         self.logger.experiment.add_scalars("dice", {"val_dice": mean_val_dice}, self.current_epoch)
         self.logger.log_metrics(tensorboard_logs, step=self.current_epoch)
 
@@ -207,29 +214,24 @@ class Net(L.pytorch.LightningModule):
 
 
 def main():
-    parser = argparse.ArgumentParser(description="COVID-19 Segmentation")
-    parser.add_argument('--architecture', type=str, default='unetr', help='Model arch: unet, resnet, etc.')
-    parser.add_argument('--metrics', type=str, default='dice', help='Metrics to use: dice, iou, etc.')
-    parser.add_argument('--epochs', type=int, default=1300, help='Number of epochs to train')
-    parser.add_argument('--batch_size', type=int, default=2, help='Batch size for training')
-    parser.add_argument('--learning_rate', type=float, default=0.001, help='Learning rate for optimizer')
-    parser.add_argument('--mask_data_path', type=str, default=INFECTION_MASKS_PATH, help='Path to the dataset')
-    parser.add_argument('--verbose', type=str, help='Select output verbosity [INFO, DEBUG, ERROR, WARNING]')
-    parser.add_argument('--seed', type=int, default=SEED, help='Seed for reproducibility')
-    parser.add_argument('--checkpoint', type=int, help='Path to model checkpoint')
-    config = parser.parse_args()
+    # Initialize the configuration
+    config = Config()
+    config.set_logging_verbose()
 
+    # Set the seed for reproducibility
     L.seed_everything(SEED)
 
     net = Net()
+    net.config = config
+    net.configure_training_process()
 
     tensorboard_logger = (L.pytorch.loggers.TensorBoardLogger(save_dir="lightning_logs", name="lightning_logs"))
     callbacks = [L.pytorch.callbacks.ModelCheckpoint(monitor="val_loss", save_top_k=1, mode="min")]
 
     trainer = L.pytorch.Trainer(
-        devices=[0],
-        accelerator="gpu",
-        max_epochs=config.epochs,
+        devices="auto",
+        accelerator="auto",
+        max_epochs=config.get_epochs(),
         logger=tensorboard_logger,
         callbacks=callbacks,
         log_every_n_steps=14,
@@ -237,9 +239,9 @@ def main():
         num_sanity_val_steps=0,
     )
 
-    trainer.fit(net, ckpt_path=config.checkpoint)
+    trainer.fit(net, ckpt_path=config.get_checkpoint())
 
-    print(f"Train completed, best_metric: {net.best_val_dice:.4f} " f"at epoch {net.best_val_epoch}")
+    print(f"Train completed, best_metric: {net.best_val_dice:.4f} at epoch {net.best_val_epoch}")
 
     trainer.test(dataloaders=net.test_dataloader(), ckpt_path="best")