+++ nets/nets.py
@@ -5,9 +5,10 @@ covid_unet = UNet(
             spatial_dims=3,
             in_channels=1,
             out_channels=1,
-            channels=(16, 32, 64, 128, 256),
-            strides=(2, 2, 2, 2),
-            num_res_units=2
+            channels=(16, 32, 64, 128),
+            strides=(2, 2, 2),
+            num_res_units=2,
+            dropout=0.2
         )
 
 covid_unetr = UNETR(